+++ src/pages/ProductOrder.tsx
@@ -1,4 +1,4 @@
-import { useForm, FormProvider } from 'react-hook-form';
+import { FormProvider } from 'react-hook-form';
 import NavigationBar from '@/common/NavigationBar';
 import GiftCardSelector from '@/components/ProductOrder/GiftCardSelector';
 import SenderInfoSection from '@/components/ProductOrder/SenderInfoSection';
@@ -6,44 +6,11 @@ import ReceiverInfoSection from '@/components/ProductOrder/ReceiverInfo/Receiver
 import ProductInfo from '@/components/giftHome/GiftThemes/ProductInfo';
 import OrderBtn from '@/components/ProductOrder/OrderBtn';
 import styled from '@emotion/styled';
-import { useLocation } from 'react-router-dom';
-
-interface OrderFormInputs {
-  senderName: string;
-  receiverName: string;
-  receiverPhone: string;
-  quantity: number;
-}
+import { useOrderForm } from '@/hooks/useOrderForm';
 
 const ProductOrder = () => {
-  const location = useLocation();
-  const { imageURL, name, price, brandInfo } = location.state || {};
-
-  const methods = useForm<OrderFormInputs>({
-    defaultValues: {
-      senderName: '',
-      receiverName: '',
-      receiverPhone: '',
-      quantity: 1,
-    },
-  });
-
-  const { handleSubmit, setError } = methods;
-
-  const validatePhone = (phone: string): boolean => /^010\d{8}$/.test(phone);
-
-  const onSubmit = (data: OrderFormInputs) => {
-    let hasError = false;
-
-    if (!validatePhone(data.receiverPhone)) {
-      setError('receiverPhone', { message: '전화번호를 입력해주세요.' });
-      hasError = true;
-    }
-
-    if (!hasError) {
-      console.log('주문 처리', data);
-    }
-  };
+  const { imageURL, name, price, brandInfo, methods, handleSubmit } =
+    useOrderForm();
 
   return (
     <FormProvider {...methods}>
@@ -59,7 +26,7 @@ const ProductOrder = () => {
             price={price}
             brandInfo={brandInfo}
           />
-          <OrderBtn price={price.basicPrice} onClick={handleSubmit(onSubmit)} />
+          <OrderBtn price={price.basicPrice} onClick={handleSubmit} />
         </Content>
       </Layout>
     </FormProvider>
